requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
et-xmlfile==1.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
multidict==6.0.4
numpy==1.24.2
openai==0.18.1
OpenAIAuth==0.3.2
openpyxl==3.1.2
pandas==2.0.0
pandas-stubs==2.0.0.230412
prompt-toolkit==3.0.38
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
requests==2.28.2
revChatGPT==3.3.5
six==1.16.0
slack==0.0.2
slack-bolt==1.17.2
slack-sdk==3.21.2
sniffio==1.3.0
socksio==1.0.0
tiktoken==0.3.3
tqdm==4.65.0
types-pytz==2023.3.0.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.8.2
